Add PNG download button to QR generator
